services/scout/requirements.txt
annotated-types==0.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
certifi==2025.1.31
cloud-sql-python-connector==1.18.1
cryptography==44.0.2
dnspython==2.7.0
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-geo-type==0.3.11
google-maps-places==0.1.24
google-crc32c==1.7.1
googleapis-common-protos==1.69.1
greenlet==3.1.1
google-resumable-media==2.7.2
idna==3.10
proto-plus==1.26.1
pg8000==1.31.2
protobuf==5.29.3
psycopg2-binary==2.9.10
pydantic==2.11.1
python-dateutil==2.9.0.post0
pydantic_core==2.33.0
requests==2.32.3
scramp==1.4.5
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0